refactor: coord out of game
