+++ src/game/mod.rs
@@ -1,1 +1,3 @@
+pub mod board;
+pub mod coord;
 pub mod game;
