+++ src/pieces/mod.rs
@@ -1,9 +1,9 @@
 use std::cmp::Ordering;
 
-use crate::game::game::{Coord, GameBoard};
-
 use self::{bishop::Bishop, king::King, knight::Knight, pawn::Pawn, queen::Queen, rook::Rook};
 use super::constants::DisplayMode;
+use crate::game::coord::Coord;
+use crate::game::game::GameBoard;
 
 pub mod bishop;
 pub mod king;
