+++ src/pieces/queen.rs
@@ -1,7 +1,8 @@
 use super::rook::Rook;
 use super::{Movable, PieceColor, PieceMove, Position};
-use crate::game::game::{Coord, GameBoard};
 use crate::constants::DisplayMode;
+use crate::game::coord::Coord;
+use crate::game::game::GameBoard;
 use crate::pieces::bishop::Bishop;
 use crate::utils::{cleaned_positions, impossible_positions_king_checked};
